refactor: 레거시 파티 코스 수정 페이지 정리
