--- src/pages/CourseSuggestPage/AddPlanBtn/index.jsx
@@ -1,11 +1,0 @@
-function AddPlanBtn() {
-  return (
-    <div className="flex justify-center my-20">
-      <button className="mx-auto h-12 text-white rounded-full text-lg w-64 md:w-80 bg-primary">
-        여행 일정에 추가하기
-      </button>
-    </div>
-  );
-}
-
-export default AddPlanBtn;
