--- src/pages/CourseSuggestPage/Atoms/PlaceDetail/index.jsx
@@ -1,9 +1,0 @@
-function PlaceDetail() {
-  return (
-    <div>
-      <div>상세설명</div>
-    </div>
-  );
-}
-
-export default PlaceDetail;
