+++ src/pages/DestinationPage/PlaceInfoBox/PlaceInfoTitle/index.jsx
@@ -5,10 +5,10 @@ function PlaceInfoTitle({ name, views, avgRate, address }) {
     <div className="mb-4">
       <div className="mb-1 pt-6 flex">
         <p className="text-2xl font-bold">{name}</p>
-        <div className="flex gap-1 m-2 mt-3 text-xs">
+        <div className="flex gap-1 items-center m-2 mt-3 text-xs">
           <span>{`방문 ${views}회`}</span>
           <span>|</span>
-          <img className="mb-1" src={star} />
+          <img src={star} />
           <span>{avgRate ? avgRate.toFixed(1) : "0.0"}</span>
         </div>
       </div>
