--- src/pages/EditCoursePage/Platform/index.jsx
@@ -1,10 +1,0 @@
-function Platform() {
-  return (
-    <div className="py-6">
-      <p className="text-2xl font-bold">플랫폼 수수료</p>
-      <p className="text-xl text-darkgray">60,000원</p>
-    </div>
-  );
-}
-
-export default Platform;
