+++ src/pages/NewPartyPage/Reservation/TravelerBox/TravelerInput/index.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { makePhoneNumber } from "../../../../utils";
+import { makePhoneNumber } from "../../../../../utils";
 
 function TravelerInput({ index }) {
   const [name, setName] = useState("");
