+++ src/pages/NewPartyPage/Reservation/index.jsx
@@ -1,16 +1,16 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { postNewParty } from "../../../api/party";
 import PartyImageBox from "../../../components/PartyImageBox";
-import PartyNumberBox from "../../CourseSuggestPage/PartyNumberBox";
+import PartyNumberBox from "./PartyNumberBox";
 import ToTalPrice from "../../PartyPage/ToTalPrice";
 import FirstCredit from "../../../components/FirstCredit";
 import SecondCredit from "../../../components/SecondCredit";
-import TravelerBox from "../../CourseSuggestPage/TravelerBox";
-import TravelerGreet from "../../CourseSuggestPage/Atoms/TravelerGreet";
+import TravelerBox from "./TravelerBox";
+import TravelerGreet from "./TravelerGreet";
 import PartyPlan from "../../../components/PartyPlan";
 import Credit from "../../../components/Credit";
-import Agreement from "../../CourseSuggestPage/AddAgree";
+import Agreement from "./AddAgree";
 import BottomRefund from "../../../components/BottomRefund";
 import CheckModal from "../../../components/CheckModal";
 import Loading from "../../../components/Loading";
